move to independent class
